I complete changed the signature and structure of this project.
I realized that a lot of things were either unnecessary or pre-optimizations that aren't need at this time of the projects life. At the current moment, I need simple solutions and a simple file strucuture. All of the important functions have been moved to __init__.py, FastClickStrategy has been deleted, and SupportsClick Protocol is in the same file as BasicClickStrategy.