got test working again, and did another big change to the api.
I complete removed main.py module and put its contents into __init.__py. This really helps reduce issues with patching, and it doesn't make sense for the main cli runners to be separate from the package itsself. If this was a really package, no one would run the main module by itsself, it would be with the cli and the rest of the app.